requirements.txt
absl-py==0.8.1
aniso8601==8.0.0
asn1crypto==1.2.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.1.0
certifi==2021.5.30
cffi==1.13.2
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
constantly==15.1.0
cryptography==2.3.1
cryptography-vectors==2.3.1
cssselect==1.1.0
cycler==0.10.0
cymem==1.31.2
Cython==0.29.24
cytoolz==0.8.2
dask==2.8.1
dataclasses==0.8
deap==1.3.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.9
entrypoints==0.3
fastcache==1.1.0
feather-format==0.4.0
Flask==1.1.1
Flask-RESTful==0.3.7
gast==0.2.2
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.14.1
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
imgaug==0.4.0
importlib-metadata==0.23
imutils==0.5.3
incremental==17.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.4.3
murmurhash==0.28.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.3
olefile==0.46
open3d==0.8.0.0
opencv-python==4.3.0.36
opt-einsum==0+untagged.49.gdbede45.dirty
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
plac==0.9.6
pluggy==0.13.1
preshed==1.0.1
primesense==2.2.0.30.post5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.1
py==1.8.0
pyarrow==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.4
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.4.4
pynput==1.7.6
pyOpenSSL==19.0.0
pyparsing==2.4.5
pypylon==1.7.2
pyreadline==2.1
pyrsistent==0.15.6
pytest==5.3.0
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.6.0
queuelib==1.5.0
regex==2018.6.21
requests==2.13.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.6.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.6.4.post2
sip==4.19.13
six==1.13.0
soupsieve==2.2
spacy==2.0.11
statsmodels==0.10.1
stopit==1.1.2
sympy==1.4
tensorboard==2.0.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==6.10.1
toolz==0.10.0
torch==1.10.2
torchvision==0.1.6
tornado==6.0.3
TPOT==0.11.0
tqdm==4.39.0
traitlets==4.3.3
Twisted==19.10.0
typed-ast==1.4.0
typing-extensions==3.10.0.2
ujson==1.35
update-checker==0.16
utils-misc==0.0.6
vtk==9.0.1
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
